Patch applied from bug #1664
Monotone-Parent: 1e79eab4c546452715e5171f1c326b7d36f1c195
Monotone-Revision: 99dc7d9dce88b8a323b1e7aab60de99ed2bfda60

Monotone-Author: ludovic@Sophos.ca
Monotone-Date: 2012-05-29T18:56:15

+++ SOPE/NGCards/CardGroup.m
@@ -354,7 +354,7 @@ static NGCardsSaxHandler *sax = nil;
 
 - (void) setChildrenAsCopy: (NSMutableArray *) someChildren
 {
-  unsigned int count, max;
+  NSUInteger count, max;
 
   ASSIGN (children, someChildren);
 
@@ -399,7 +399,7 @@ static NGCardsSaxHandler *sax = nil;
 - (NSString *) description
 {
   NSMutableString *str;
-  unsigned int count, max;
+  NSUInteger count, max;
 
   str = [NSMutableString stringWithCapacity:64];
   [str appendFormat:@"<%p[%@]:%@",
@@ -421,7 +421,7 @@ static NGCardsSaxHandler *sax = nil;
 - (void) replaceThisElement: (CardElement *) oldElement
                 withThisOne: (CardElement *) newElement
 {
-  unsigned int index;
+  NSUInteger index;
 
   index = [children indexOfObject: oldElement];
   if (index != NSNotFound)
